+++ tests/snapshot_tests/test_snapshots.py
@@ -1340,13 +1340,13 @@ def hello(name):
         text_area = pilot.app.query_one(TextArea)
         text_area.load_text(text)
         text_area.language = "python"
-        text_area.selection = Selection((0, 1), (1, 9))
         text_area.theme = theme_name
 
     assert snap_compare(
         SNAPSHOT_APPS_DIR / "text_area.py",
         run_before=setup_theme,
         terminal_size=(48, text.count("\n") + 4),
+        press=['right', 'shift+down'] + ['shift+right'] * 8,
     )
 
 
@@ -1878,7 +1878,7 @@ def test_maximize_container(snap_compare):
         DEFAULT_CSS = """
         FormContainer {
             width: 50%;
-            border: blue;            
+            border: blue;
         }
         """
 
@@ -2577,7 +2577,7 @@ def test_themes(snap_compare, theme_name):
         Screen {
             align: center middle;
         }
-        
+
         Label {
             background: $panel;
             color: $text;
@@ -2612,7 +2612,7 @@ def test_custom_theme_with_variables(snap_compare):
         Screen {
             align: center middle;
         }
-        
+
         Label {
             background: $custom-background;
             color: $custom-text;
@@ -2787,7 +2787,7 @@ def test_position_absolute(snap_compare):
 
     class AbsoluteApp(App):
         CSS = """
-        Screen {        
+        Screen {
             align: center middle;
 
             .absolute {
@@ -2802,7 +2802,7 @@ def test_position_absolute(snap_compare):
                 offset: 1 1;
             }
             .offset2 {
-                offset: 2 2;                
+                offset: 2 2;
             }
             .offset3 {
                 offset: 3 3;
@@ -2840,7 +2840,7 @@ def test_grid_offset(snap_compare):
             border: solid green;
         }
 
-        #six {   
+        #six {
             offset: 0 10;
             background: blue;
         }
@@ -3094,7 +3094,7 @@ def test_dock_align(snap_compare):
             layout: horizontal;
         }
 
-        MainContainer {    
+        MainContainer {
             width: 100%;
             height: 100%;
             background: red;
@@ -3153,7 +3153,7 @@ def test_auto_parent_with_alignment(snap_compare):
                 height: auto;
                 background: red;
                 border: white;
-            }        
+            }
         }
         """
 
@@ -3239,7 +3239,7 @@ Where the fear has gone there will be nothing. Only I will remain."""
             height: auto;
             background: blue;
         }
-        Label {        
+        Label {
             border: heavy red;
             text-align: left;
         }
@@ -3292,12 +3292,12 @@ def test_collapsible_datatable(snap_compare):
     class MyApp(App):
         CSS = """
         DataTable {
-            max-height: 1fr;            
-            border: red;            
+            max-height: 1fr;
+            border: red;
         }
         Collapsible {
             max-height: 50%;
-            
+
         }
         """
 
@@ -3417,7 +3417,7 @@ def test_overflow(snap_compare):
     class OverflowApp(App):
         CSS = """
         Label {
-            max-width: 100vw;            
+            max-width: 100vw;
         }
         #label1 {
             # Overflow will be cropped
@@ -3512,7 +3512,7 @@ def test_option_list_wrapping(snap_compare):
 
     class OLApp(App):
         CSS = """
-        OptionList { 
+        OptionList {
             width: 40;
             text-wrap: nowrap;
             text-overflow: ellipsis;
@@ -3823,7 +3823,7 @@ def test_select_list_in_collapsible(snap_compare):
     class CustomWidget(Horizontal):
         DEFAULT_CSS = """
         CustomWidget {
-            height: auto;                	
+            height: auto;
         }
         """